fix CRAN check issues

+++ src/cpglmm.c
@@ -270,7 +270,7 @@ static double cp_update_L(SEXP x)
     // compute the cholesky factor of AA'+I with permutation
     if (!M_cholmod_factorize_p(A, one, (int*)NULL, 0, L, &c))
 	error(_("cholmod_factorize_p failed: status %d, minor %d from ncol %d"),
-	      c.status, L->minor, L->n);
+	      (int)c.status, (int)L->minor, (int)L->n);
 
     
     d[ldL2_POS] = M_chm_factor_ldetL2(L);
@@ -1093,7 +1093,7 @@ SEXP mer_create_L(SEXP CmP)
     L = M_cholmod_analyze(Cm, &c);
     if (!M_cholmod_factorize_p(Cm, one, (int*)NULL, 0 /*fsize*/, L, &c))
 	error(_("cholmod_factorize_p failed: status %d, minor %d from ncol %d"),
-	      c.status, L->minor, L->n);
+	      (int)c.status, (int)L->minor, (int)L->n);
 
     return M_chm_factor_to_SEXP(L, 1);
 }
